refactor!: remove completions
I'm removing completions because they don't work when this program is
invoked as a subcommand of `gh`. If you run it directly to work around
this limitation as the README previously suggested, then even though
completions work, the user won't be properly authenticated. This could
lead to unexpected behaviour, so it shouldn't be encouraged.